+++ src/pages/administracao-de-marketing/p.tsx
@@ -56,7 +56,6 @@ function Page() {
 
       <ProductDetails product={product} />
 
-
       <PDPProductShelf
         pretitle=""
         title={`Mais vendidos de ${departmentName}`}