+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **GM-132** - Contact us form
+
 ## [0.10.0] - 2022-08-04
 
 ### Added