+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **GM-109** - Add component product description
 - **GM-137** - Add static institutional pages
 - **GM-131** - Add institutional menu
+
 ## [0.10.0] - 2022-08-04
 
 ### Added